backend/requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0

# Optimization
optimum[onnxruntime]>=1.12.0

# Data & Metrics
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
ftfy>=6.1.1
alt-profanity-check>=1.7.2

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0